moved logic to investments_strategy file
